Moved config class.
