+++ src/main/java/org/hobbit/core/components/AbstractDataGenerator.java
@@ -24,7 +24,7 @@ import org.hobbit.core.Commands;
 import org.hobbit.core.Constants;
 import org.hobbit.core.rabbit.DataSender;
 import org.hobbit.core.rabbit.DataSenderImpl;
-import org.hobbit.utils.HobbitConfiguration;
+import org.hobbit.utils.config.HobbitConfiguration;
 
 public abstract class AbstractDataGenerator extends AbstractPlatformConnectorComponent {
