+++ src/test/java/org/hobbit/core/components/ContainerEnvironmentTest.java
@@ -22,9 +22,9 @@ import java.util.Collection;
 import org.apache.commons.configuration2.Configuration;
 import org.apache.commons.configuration2.PropertiesConfiguration;
 import org.hobbit.core.components.dummy.DummyCommandReceivingComponent;
+import org.hobbit.utils.config.HobbitConfiguration;
 import org.hobbit.core.Constants;
 import org.hobbit.core.TestConstants;
-import org.hobbit.utils.HobbitConfiguration;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.runner.RunWith;
