+++ src/test/java/org/hobbit/core/components/dummy/DummyCommandReceivingComponent.java
@@ -17,8 +17,7 @@
 package org.hobbit.core.components.dummy;
 
 import org.hobbit.core.components.AbstractCommandReceivingComponent;
-
-import org.hobbit.utils.HobbitConfiguration;
+import org.hobbit.utils.config.HobbitConfiguration;
 import org.junit.Ignore;
 
 @Ignore
