+++ src/test/java/org/hobbit/core/components/dummy/DummyEvalStoreReceiver.java
@@ -24,7 +24,7 @@ import java.util.List;
 import org.hobbit.core.components.AbstractEvaluationStorage;
 import org.hobbit.core.data.ResultPair;
 import org.hobbit.core.rabbit.RabbitMQUtils;
-import org.hobbit.utils.HobbitConfiguration;
+import org.hobbit.utils.config.HobbitConfiguration;
 import org.junit.Ignore;
 
 @Ignore
