+++ src/test/java/org/hobbit/core/components/dummy/DummySystemReceiver.java
@@ -23,7 +23,7 @@ import java.util.List;
 
 import org.hobbit.core.components.AbstractSystemAdapter;
 import org.hobbit.core.rabbit.RabbitMQUtils;
-import org.hobbit.utils.HobbitConfiguration;
+import org.hobbit.utils.config.HobbitConfiguration;
 import org.junit.Ignore;
 
 @Ignore
