+++ src/test/java/org/hobbit/core/components/dummy/DummyTaskGenReceiver.java
@@ -22,7 +22,7 @@ import java.util.List;
 
 import org.hobbit.core.components.AbstractTaskGenerator;
 import org.hobbit.core.rabbit.RabbitMQUtils;
-import org.hobbit.utils.HobbitConfiguration;
+import org.hobbit.utils.config.HobbitConfiguration;
 import org.junit.Ignore;
 
 @Ignore
